// packages/core/src/schemas/dml-picture.ts
/**
 * Auto-generated from dml-picture.xsd
 * DO NOT EDIT - Generated by xsd-converter
 */

import type { XsdSchema, XsdSimpleType, XsdComplexType, XsdElement, XsdGroup, XsdAttributeGroup } from '../types'

export const dmlPictureSchema: XsdSchema = {
  "targetNamespace": "http://purl.oclc.org/ooxml/drawingml/picture",
  "imports": [
    {
      "namespace": "http://purl.oclc.org/ooxml/drawingml/main",
      "schemaLocation": "dml-main.xsd"
    }
  ],
  "simpleTypes": [],
  "complexTypes": [
    {
      "name": "CT_PictureNonVisual",
      "mixed": false,
      "abstract": false,
      "attributes": [],
      "attributeGroupRefs": [],
      "content": {
        "kind": "elementOnly",
        "compositor": {
          "kind": "sequence",
          "particles": [
            {
              "type": "a:CT_NonVisualDrawingProps",
              "name": "cNvPr",
              "minOccurs": 1,
              "maxOccurs": 1,
              "nillable": false
            },
            {
              "type": "a:CT_NonVisualPictureProperties",
              "name": "cNvPicPr",
              "minOccurs": 1,
              "maxOccurs": 1,
              "nillable": false
            }
          ],
          "minOccurs": 1,
          "maxOccurs": 1
        }
      }
    },
    {
      "name": "CT_Picture",
      "mixed": false,
      "abstract": false,
      "attributes": [],
      "attributeGroupRefs": [],
      "content": {
        "kind": "elementOnly",
        "compositor": {
          "kind": "sequence",
          "particles": [
            {
              "type": "CT_PictureNonVisual",
              "name": "nvPicPr",
              "minOccurs": 1,
              "maxOccurs": 1,
              "nillable": false
            },
            {
              "type": "a:CT_BlipFillProperties",
              "name": "blipFill",
              "minOccurs": 1,
              "maxOccurs": 1,
              "nillable": false
            },
            {
              "type": "a:CT_ShapeProperties",
              "name": "spPr",
              "minOccurs": 1,
              "maxOccurs": 1,
              "nillable": false
            }
          ],
          "minOccurs": 1,
          "maxOccurs": 1
        }
      }
    }
  ],
  "elements": [
    {
      "name": "pic",
      "type": {
        "kind": "local",
        "name": "CT_Picture"
      },
      "minOccurs": 1,
      "maxOccurs": 1,
      "nillable": false
    }
  ],
  "groups": [],
  "attributeGroups": []
}

export default dmlPictureSchema